refactor(twitter): share user timeline transport
