+++ clis/twitter/tweets.js
@@ -1,155 +1,20 @@
 import { cli, Strategy } from '@jackwener/opencli/registry';
-import { ArgumentError, AuthRequiredError, CommandExecutionError, EmptyResultError } from '@jackwener/opencli/errors';
-import { resolveTwitterOperationMetadata, sanitizeQueryId, extractMedia, extractQuotedTweet, normalizeTwitterGraphqlPayload, unwrapBrowserResult, normalizeTwitterScreenName, describeTwitterApiError } from './shared.js';
-import { TWITTER_BEARER_TOKEN, applyTopByEngagement } from './utils.js';
+import { ArgumentError, CommandExecutionError, EmptyResultError } from '@jackwener/opencli/errors';
+import { sanitizeQueryId, extractMedia, extractQuotedTweet, describeTwitterApiError } from './shared.js';
+import { applyTopByEngagement } from './utils.js';
+import {
+    MAX_USER_TWEETS_PAGES,
+    USER_TWEETS_PAGE_SIZE,
+    MAX_USER_TWEETS_LIMIT,
+    DEFAULT_USER_TWEETS_PAGE_DELAY_SECONDS,
+    buildUserTweetsUrl,
+    buildUserByScreenNameUrl,
+    fetchUserTimelinePage,
+    resolveUserTimelineContext,
+} from './user-timeline.js';
 
-const USER_TWEETS_QUERY_ID = 'lrMzG9qPQHpqJdP3AbM-bQ';
-const USER_BY_SCREEN_NAME_QUERY_ID = 'IGgvgiOx4QZndDHuD3x9TQ';
-const MAX_PAGINATION_PAGES = 100;
-const USER_TWEETS_PAGE_SIZE = 100;
-const MAX_TWEETS_LIMIT = MAX_PAGINATION_PAGES * USER_TWEETS_PAGE_SIZE;
-const DEFAULT_PAGE_DELAY_SECONDS = 2;
-
-const USER_TWEETS_FEATURES = {
-    rweb_video_screen_enabled: true,
-    rweb_cashtags_enabled: true,
-    payments_enabled: false,
-    profile_label_improvements_pcf_label_in_post_enabled: true,
-    responsive_web_profile_redirect_enabled: true,
-    rweb_tipjar_consumption_enabled: true,
-    verified_phone_label_enabled: false,
-    creator_subscriptions_tweet_preview_api_enabled: true,
-    responsive_web_graphql_timeline_navigation_enabled: true,
-    responsive_web_graphql_skip_user_profile_image_extensions_enabled: false,
-    premium_content_api_read_enabled: false,
-    communities_web_enable_tweet_community_results_fetch: true,
-    c9s_tweet_anatomy_moderator_badge_enabled: true,
-    responsive_web_grok_analyze_button_fetch_trends_enabled: false,
-    responsive_web_grok_analyze_post_followups_enabled: true,
-    rweb_cashtags_composer_attachment_enabled: true,
-    responsive_web_jetfuel_frame: true,
-    responsive_web_grok_share_attachment_enabled: true,
-    responsive_web_grok_annotations_enabled: true,
-    articles_preview_enabled: true,
-    responsive_web_edit_tweet_api_enabled: true,
-    graphql_is_translatable_rweb_tweet_is_translatable_enabled: true,
-    view_counts_everywhere_api_enabled: true,
-    longform_notetweets_consumption_enabled: true,
-    responsive_web_twitter_article_tweet_consumption_enabled: true,
-    tweet_awards_web_tipping_enabled: false,
-    content_disclosure_indicator_enabled: true,
-    content_disclosure_ai_generated_indicator_enabled: true,
-    responsive_web_grok_show_grok_translated_post: false,
-    responsive_web_grok_analysis_button_from_backend: true,
-    post_ctas_fetch_enabled: false,
-    freedom_of_speech_not_reach_fetch_enabled: true,
-    standardized_nudges_misinfo: true,
-    tweet_with_visibility_results_prefer_gql_limited_actions_policy_enabled: true,
-    longform_notetweets_rich_text_read_enabled: true,
-    longform_notetweets_inline_media_enabled: true,
-    responsive_web_grok_image_annotation_enabled: true,
-    responsive_web_grok_imagine_annotation_enabled: true,
-    responsive_web_grok_community_note_auto_translation_is_enabled: false,
-    responsive_web_enhance_cards_enabled: false,
-};
-
-const USER_TWEETS_FIELD_TOGGLES = {
-    withPayments: true,
-    withAuxiliaryUserLabels: true,
-    withArticleRichContentState: true,
-    withArticlePlainText: true,
-    withArticleSummaryText: true,
-    withArticleVoiceOver: true,
-    withGrokAnalyze: true,
-    withDisallowedReplyControls: true,
-};
-
-const USER_BY_SCREEN_NAME_FEATURES = {
-    hidden_profile_subscriptions_enabled: true,
-    profile_label_improvements_pcf_label_in_post_enabled: true,
-    responsive_web_profile_redirect_enabled: true,
-    rweb_tipjar_consumption_enabled: true,
-    responsive_web_graphql_exclude_directive_enabled: true,
-    verified_phone_label_enabled: false,
-    subscriptions_verification_info_is_identity_verified_enabled: true,
-    subscriptions_verification_info_verified_since_enabled: true,
-    highlights_tweets_tab_ui_enabled: true,
-    responsive_web_twitter_article_notes_tab_enabled: true,
-    subscriptions_feature_can_gift_premium: true,
-    creator_subscriptions_tweet_preview_api_enabled: true,
-    responsive_web_graphql_skip_user_profile_image_extensions_enabled: false,
-    responsive_web_graphql_timeline_navigation_enabled: true,
-};
-
-const USER_BY_SCREEN_NAME_FIELD_TOGGLES = {
-    withPayments: true,
-    withAuxiliaryUserLabels: true,
-};
-
-const USER_TWEETS_OPERATION = {
-    queryId: USER_TWEETS_QUERY_ID,
-    features: USER_TWEETS_FEATURES,
-    fieldToggles: USER_TWEETS_FIELD_TOGGLES,
-};
-
-const USER_BY_SCREEN_NAME_OPERATION = {
-    queryId: USER_BY_SCREEN_NAME_QUERY_ID,
-    features: USER_BY_SCREEN_NAME_FEATURES,
-    fieldToggles: USER_BY_SCREEN_NAME_FIELD_TOGGLES,
-};
-
-function normalizeUserTweetsOperation(operation) {
-    if (typeof operation === 'string') {
-        return { queryId: operation, features: USER_TWEETS_FEATURES, fieldToggles: USER_TWEETS_FIELD_TOGGLES };
-    }
-    return {
-        queryId: operation?.queryId || USER_TWEETS_QUERY_ID,
-        features: operation?.features || USER_TWEETS_FEATURES,
-        fieldToggles: operation?.fieldToggles || USER_TWEETS_FIELD_TOGGLES,
-    };
-}
-
-function normalizeUserByScreenNameOperation(operation) {
-    if (typeof operation === 'string') {
-        return { queryId: operation, features: USER_BY_SCREEN_NAME_FEATURES, fieldToggles: USER_BY_SCREEN_NAME_FIELD_TOGGLES };
-    }
-    return {
-        queryId: operation?.queryId || USER_BY_SCREEN_NAME_QUERY_ID,
-        features: operation?.features || USER_BY_SCREEN_NAME_FEATURES,
-        fieldToggles: operation?.fieldToggles || USER_BY_SCREEN_NAME_FIELD_TOGGLES,
-    };
-}
-
-function appendGraphqlParams(path, variables, operation) {
-    const fieldToggles = operation.fieldToggles || {};
-    const params = [
-        `variables=${encodeURIComponent(JSON.stringify(variables))}`,
-        `features=${encodeURIComponent(JSON.stringify(operation.features || {}))}`,
-    ];
-    if (Object.keys(fieldToggles).length > 0) {
-        params.push(`fieldToggles=${encodeURIComponent(JSON.stringify(fieldToggles))}`);
-    }
-    return `${path}?${params.join('&')}`;
-}
-
-function buildUserTweetsUrl(operation, userId, count, cursor) {
-    const normalized = normalizeUserTweetsOperation(operation);
-    const vars = {
-        userId,
-        count,
-        includePromotedContent: false,
-        withQuickPromoteEligibilityTweetFields: true,
-        withVoice: true,
-    };
-    if (cursor) vars.cursor = cursor;
-    return appendGraphqlParams(`/i/api/graphql/${normalized.queryId}/UserTweets`, vars, normalized);
-}
-
-function buildUserByScreenNameUrl(operation, screenName) {
-    const normalized = normalizeUserByScreenNameOperation(operation);
-    const vars = { screen_name: screenName, withSafetyModeUserFields: true };
-    return appendGraphqlParams(`/i/api/graphql/${normalized.queryId}/UserByScreenName`, vars, normalized);
-}
+const MAX_TWEETS_LIMIT = MAX_USER_TWEETS_LIMIT;
+const DEFAULT_PAGE_DELAY_SECONDS = DEFAULT_USER_TWEETS_PAGE_DELAY_SECONDS;
 
 function extractTweet(result, seen) {
     if (!result) return null;
@@ -221,7 +86,7 @@ function normalizeLimit(rawLimit) {
     if (!Number.isInteger(limit) || limit < 1 || limit > MAX_TWEETS_LIMIT) {
         throw new ArgumentError(
             `twitter tweets --limit must be an integer between 1 and ${MAX_TWEETS_LIMIT}`,
-            `Example: opencli twitter tweets @jack --limit 250`
+            'Example: opencli twitter tweets @jack --limit 250',
         );
     }
     return limit;
@@ -232,7 +97,7 @@ function normalizePageDelaySeconds(rawDelay) {
     if (!Number.isInteger(delay) || delay < 0 || delay > 60) {
         throw new ArgumentError(
             'twitter tweets --page-delay must be an integer between 0 and 60 seconds',
-            'Example: opencli twitter tweets @jack --limit 250 --page-delay 2'
+            'Example: opencli twitter tweets @jack --limit 250 --page-delay 2',
         );
     }
     return delay;
@@ -256,69 +121,18 @@ cli({
     func: async (page, kwargs) => {
         const limit = normalizeLimit(kwargs.limit);
         const pageDelaySeconds = normalizePageDelaySeconds(kwargs['page-delay']);
-        const rawUsername = String(kwargs.username ?? '').trim();
-        let username = normalizeTwitterScreenName(rawUsername);
-        if (rawUsername && !username) {
-            throw new ArgumentError('twitter tweets username must be a valid Twitter/X handle', 'Example: opencli twitter tweets @jack --limit 20');
-        }
-        // When no username is given, detect the logged-in user (own tweets).
-        // Mirrors the self-detection pattern used by twitter/profile and
-        // twitter/likes so agents can pull own-account data without having
-        // to know their own screen name up front.
-        if (!username) {
-            await page.goto('https://x.com/home');
-            await page.wait({ selector: '[data-testid="primaryColumn"]' });
-            // Bridge wraps primitive page.evaluate returns as { session, data:<value> }.
-            // unwrapBrowserResult drops that envelope so the href string is usable.
-            const href = unwrapBrowserResult(await page.evaluate(`() => {
-        const link = document.querySelector('a[data-testid="AppTabBar_Profile_Link"]');
-        return link ? link.getAttribute('href') : null;
-      }`));
-            if (!href || typeof href !== 'string')
-                throw new AuthRequiredError('x.com', 'Could not detect logged-in user. Are you logged in?');
-            username = normalizeTwitterScreenName(href);
-            if (!username) {
-                throw new AuthRequiredError('x.com', 'Could not detect logged-in user. Are you logged in?');
-            }
-        }
-
-        const cookies = await page.getCookies({ url: 'https://x.com' });
-        const ct0 = cookies.find((c) => c.name === 'ct0')?.value || null;
-        if (!ct0) throw new AuthRequiredError('x.com', 'Not logged into x.com (no ct0 cookie)');
-
-        const userTweetsOperation = await resolveTwitterOperationMetadata(page, 'UserTweets', USER_TWEETS_OPERATION);
-        const userByScreenNameOperation = await resolveTwitterOperationMetadata(page, 'UserByScreenName', USER_BY_SCREEN_NAME_OPERATION);
-
-        const headers = JSON.stringify({
-            'Authorization': `Bearer ${decodeURIComponent(TWITTER_BEARER_TOKEN)}`,
-            'X-Csrf-Token': ct0,
-            'X-Twitter-Auth-Type': 'OAuth2Session',
-            'X-Twitter-Active-User': 'yes',
-        });
-
-        const ubsUrl = buildUserByScreenNameUrl(userByScreenNameOperation, username);
-        const userId = unwrapBrowserResult(await page.evaluate(`async () => {
-      const resp = await fetch("${ubsUrl}", { headers: ${headers}, credentials: 'include' });
-      if (!resp.ok) return null;
-      const d = await resp.json();
-      return d?.data?.user?.result?.rest_id || null;
-    }`));
-        if (!userId) throw new CommandExecutionError(`Could not resolve @${username}`);
-
+        const context = await resolveUserTimelineContext(page, kwargs.username, { allowLoggedInDefault: true });
+        const { username } = context;
         const seen = new Set();
         const all = [];
         let cursor = null;
         // Runaway guard only; --limit and cursor exhaustion control normal pagination.
-        for (let i = 0; i < MAX_PAGINATION_PAGES && all.length < limit; i++) {
+        for (let i = 0; i < MAX_USER_TWEETS_PAGES && all.length < limit; i++) {
             if (i > 0 && pageDelaySeconds > 0) {
                 await page.wait(pageDelaySeconds);
             }
             const fetchCount = Math.min(USER_TWEETS_PAGE_SIZE, limit - all.length + 10);
-            const url = buildUserTweetsUrl(userTweetsOperation, userId, fetchCount, cursor);
-            const data = normalizeTwitterGraphqlPayload(await page.evaluate(`async () => {
-        const r = await fetch("${url}", { headers: ${headers}, credentials: 'include' });
-        return r.ok ? await r.json() : { error: r.status };
-      }`));
+            const data = await fetchUserTimelinePage(page, context, cursor, fetchCount);
             if (data?.error) {
                 if (all.length === 0) throw new CommandExecutionError(describeTwitterApiError('UserTweets', data.error));
                 break;
@@ -328,10 +142,8 @@ cli({
             if (!nextCursor || nextCursor === cursor) break;
             cursor = nextCursor;
         }
-
         if (all.length === 0) throw new EmptyResultError(`@${username} has no recent tweets`, 'Account may be private or suspended');
-        const trimmed = all.slice(0, limit);
-        return applyTopByEngagement(trimmed, kwargs['top-by-engagement']);
+        return applyTopByEngagement(all.slice(0, limit), kwargs['top-by-engagement']);
     },
 });
